added basic serialization
added a feature to the create called "serialize" that adds the ability to save and load the actions  of an input map to a file. also made an example of how one might use said feature and a bunch of tests to make sure what is save and what is loaded is the same.

could maybe make it auto load files if the exist but this would need some thought about how to expose the save path to the developer.

could also add a version/merge system so that if an old keybind file is loaded new key binds will be added automatically; didn't do the auto merging here because it would mean working out a way to not override the custom keybinds and also not conflict with them